Agrego todas las rutas de paquetes

+++ src/controllers/paquetes.controller.js
@@ -3,8 +3,30 @@ import { Paquete } from "../models/PaqueteTuristico.js";
 
 export const GetPaquetes = async(req , res) => {
     try {
-        const paquetes = await Paquete.find();
+        const paquetes = await Paquete.find()
+        .select("-__v")
+        .lean()
+        .exec()
+
         return res.json({paquetes})
+    } catch (error) {
+        res.status(500).json({error: "Error de servidor"})
+    }
+}
+
+
+export const GetPaqueteById = async(req , res) => {
+    try {
+        const {id} = req.params;
+        const paquete = await Paquete.findById(id)
+        .select("-__v")
+        .lean()
+        .exec()
+
+        if (!paquete)
+            return res.status(404).json({ error: `No existe un paquete con ID ${id}` });
+
+        return res.json({paquete})
     } catch (error) {
         res.status(500).json({error: "Error de servidor"})
     }
@@ -15,6 +37,11 @@ export const GetPaquetes = async(req , res) => {
 export const PostPaquetes = async(req , res) => {
     try {
         const {titulo, descripcion, precio} = req.body
+
+        const existePaquete = await validarTitulo(titulo, null);
+        if(existePaquete){
+            return res.status(409).json({ error: `Ya existe un paquete con un título '${titulo}'. Por favor, ingrese un título diferente.` });
+        }
 
         const paquete = new Paquete({
             titulo,
@@ -27,6 +54,7 @@ export const PostPaquetes = async(req , res) => {
         return res.status(201).json({paquete})
 
     } catch (error) {
+        console.log(error)
         res.status(500).json({error: "Error de servidor"})
     }
 }
@@ -40,7 +68,13 @@ export const PutPaquetes = async(req , res) => {
 
         let paquete = await Paquete.findById(id)
         if (!paquete)
-            return res.status(404).json({ error: "No existe el paquete" });
+            return res.status(404).json({ error: `No existe un paquete con ID ${id}` });
+
+        
+        const existePaquete = await validarTitulo(titulo, id);
+        if(existePaquete){
+            return res.status(409).json({ error: `Ya existe un paquete con un título '${titulo}'. Por favor, ingrese un título diferente.` });
+        }
 
         paquete.titulo = titulo ?? proyecto.titulo;
         paquete.descripcion = descripcion ?? paquete.descripcion;
@@ -50,7 +84,8 @@ export const PutPaquetes = async(req , res) => {
         return res.status(200).json({paquete})
 
     } catch (error) {
-        return res.status(500).json({error: "Error de servidor"})
+        console.log(error)
+        return res.status(500).json({msg: `El paquete con ID ${id} fue modificado exitosamente`, error: "Error de servidor"})
     }
 }
 
@@ -60,10 +95,28 @@ export const DeletePaquetes = async(req , res) => {
     try {
         const {id} = req.params;
         const paquete = await Paquete.findById(id)
+        if (!paquete)
+            return res.status(404).json({ error: `No existe un paquete con ID ${id}` });
+
         await paquete.deleteOne()
-        return res.status(204).json()
+        return res.status(200).json({msg: `El paquete con ID ${id} fue eliminado exitosamente`})
 
     } catch (error) {
         return res.status(500).json({error: "Error de servidor"})
     }
+}
+
+
+const validarTitulo = async (titulo, id) => {
+    let paquete = null;
+    if(id == null){
+        paquete = await Paquete.findOne({titulo: titulo})
+    } else {
+        paquete = await Paquete.findOne({titulo: titulo, _id: {$ne:id}})
+    }
+
+    if (!paquete){
+        return false;
+    } 
+    return true;
 }
